Extract Turing tests
