+++ src/DynamicPPL.jl
@@ -29,6 +29,7 @@ import Base: Symbol,
 export  AbstractVarInfo,
         VarInfo,
         UntypedVarInfo,
+        TypedVarInfo,
         getlogp,
         setlogp!,
         acclogp!,
@@ -39,6 +40,7 @@ export  AbstractVarInfo,
         increment_num_produce!,
         set_retained_vns_del_by_spl!,
         is_flagged,
+        set_flag!,
         unset_flag!,
         setgid!,
         updategid!,
