+++ test/Project.toml
@@ -1,61 +1,27 @@
 [deps]
 AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
-AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
-AdvancedMH = "5b7e9947-ddc0-4b3f-9b55-0d8042f74170"
-AdvancedPS = "576499cb-2369-40b2-a588-c64705576edc"
-AdvancedVI = "b5ca4192-6429-45e5-a2d9-87aec30a685c"
-BangBang = "198e06fe-97b7-11e9-32a5-e1d131e6ad66"
 Bijectors = "76274a88-744f-5084-9051-94815aaf08c4"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DistributionsAD = "ced4e74d-a319-5a8a-b0ac-84af2272839c"
-DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
-EllipticalSliceSampling = "cad2338a-1db2-11e9-3401-43bc07c9ede2"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-Libtask = "6f1fad26-d15e-5dc8-ae53-837a1d7b8c9f"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-LogDensityProblems = "6fdf6af0-433a-55f7-b3ed-c6c6e0b8df7c"
 MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
-NamedArrays = "86f7a689-2022-50b4-a561-43c23ac3c673"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Serialization = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
-SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
-StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Tracker = "9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
-ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 
 [compat]
 AbstractMCMC = "2.1"
-AdvancedHMC = "0.2.25"
-AdvancedMH = "0.5.2"
-AdvancedPS = "0.1"
-AdvancedVI = "0.1"
-BangBang = "0.3"
 Bijectors = "0.8.2"
-Distributions = "0.23.8, 0.24"
+Distributions = "0.24"
 DistributionsAD = "0.6.3"
-DocStringExtensions = "0.8.2"
-EllipticalSliceSampling = "0.4"
 ForwardDiff = "0.10.12"
-Libtask = "0.4.1, 0.5"
-LogDensityProblems = "0.10.3"
 MCMCChains = "4.0.4"
 MacroTools = "0.5.5"
-NamedArrays = "0.9"
-Reexport = "0.2"
-Requires = "1.0.1"
-SpecialFunctions = "0.10.3, 1.0"
-StatsBase = "0.33"
-StatsFuns = "0.9.5"
 Tracker = "0.2.11"
 Zygote = "0.5.4, 0.6"
-ZygoteRules = "0.2"
 julia = "1.3"
