+++ test/runtests.jl
@@ -1,21 +1,23 @@
 using DynamicPPL
+using Bijectors
 using Distributions
+using DistributionsAD
 using ForwardDiff
+using MacroTools
+using MCMCChains
 using Tracker
 using Zygote
 
 using Distributed
+using Pkg
 using Random
 using Serialization
 using Test
 
-dir = splitdir(splitdir(pathof(DynamicPPL))[1])[1]
-include(dir*"/test/Turing/Turing.jl")
-using .Turing
+using DynamicPPL: vsym, vinds, getargs_dottilde, getargs_tilde, Selector
 
-setprogress!(false)
+Random.seed!(100)
 
-include("test_utils/AllUtils.jl")
 include("test_util.jl")
 
 @testset "DynamicPPL.jl" begin
@@ -28,7 +30,6 @@ include("test_util.jl")
     include("independence.jl")
     include("distribution_wrappers.jl")
     include("context_implementations.jl")
-    include("loglikelihoods.jl")
 
     include("threadsafe.jl")
 
@@ -36,5 +37,12 @@ include("test_util.jl")
 
     @testset "compat" begin
         include(joinpath("compat", "ad.jl"))
+    end
+
+    @testset "turing" begin
+        Pkg.activate("turing")
+        Pkg.develop(PackageSpec(path=dirname(@__DIR__)))
+        Pkg.instantiate()
+        include(joinpath("turing", "runtests.jl"))
     end
 end
