+++ test/turing/Project.toml
@@ -1,0 +1,9 @@
+[deps]
+DynamicPPL = "366bfd00-2699-11ea-058f-f148b4cae6d8"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
+
+[compat]
+Turing = "0.15"
+julia = "1.3"
